+++ src/test/java/com/project/post/application/service/PostLikeServiceTest.java
@@ -3,6 +3,7 @@ package com.project.post.application.service;
 import com.project.global.error.BusinessException;
 import com.project.global.error.ErrorCode;
 import com.project.post.application.dto.LikeScrapToggleResponse;
+import com.project.post.application.service.impl.PostLikeServiceImpl;
 import com.project.post.domain.repository.PostLikeRepository;
 import com.project.post.domain.repository.PostRepository;
 import com.project.user.domain.entity.User;
@@ -33,7 +34,7 @@ class PostLikeServiceTest {
     private PostLikeRepository postLikeRepository;
 
     @InjectMocks
-    private PostLikeService postLikeService;
+    private PostLikeServiceImpl postLikeService;
 
     @Test
     @DisplayName("게시글이 없으면 좋아요 추가 시 예외를 던진다")
